# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Ant API Service",
    "version": "v1",
    "description": "Ant 웹서비스를 위한 API 문서"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
    "/users/email/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Email Login process",
        "description": "이메일 로그인",
        "requestBody": {
          "description": "User data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailLogin"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "인증 성공, 토큰 반환",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "400": {
            "description": "유효성 검증이 실패하여 로그인 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "404": {
            "description": "유저를 찾을 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          }
        }
      }
    },
    "/users/email": {
      "post": {
        "tags": ["Auth"],
        "summary": "Account Create By Email",
        "description": "이메일로 계정 생성",
        "requestBody": {
          "description": "Email Account Data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinByEmail"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "이메일 계정 생성 완료",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Token"
                  }
                }
              }
            }
          },
          "400": {
            "description": "유효성 검증이 실패하여 계정 생성 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "409": {
            "description": "이미 존재하는 계정이라서 계정 생성 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserConflictError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EmailLogin": {
        "properties": {
          "email": {
            "type": "string",
            "required": "true"
          },
          "passWord": {
            "type": "string",
            "required": "true"
          }
        }
      },
      "JoinByEmail": {
        "properties": {
          "email": {
            "type": "string",
            "required": "true"
          },
          "passWord": {
            "type": "string",
            "required": "true"
          },
          "nickName": {
            "type": "string",
            "required": "true"
          }
        }
      },
      "Token": {
        "properties": {
          "token": {
            "type": "string"
          }
        }
      },
      "NotFoundError": {
        "properties": {
          "type": {
            "type": "string",
            "enum": [ "NotFoundError" ]
          },
          "details": {
            "type": "string"
          }
        }
      },
      "ValidationError": {
        "properties": {
          "type": {
            "type": "string",
            "enum": [ "ValidationError" ]
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationErrorDetail"
            }
          }
        }
      },
      "ValidationErrorDetail": {
        "properties": {
          "msg": {
            "type": "string",
            "description": "에러 원인"
          },
          "param": {
            "type": "string",
            "description": "파라미터 필드명"
          },
          "location": {
            "type": "string",
            "description": "파라미터 위치 : body, query, params, cookies 또는 headers"
          }
        }
      },
      "UserConflictError": {
        "properties": {
          "type": {
            "type": "string",
            "enum": [ "UserConflictError" ]
          },
          "details": {
            "type": "string"
          }
        }
      }
    }
  }
}